Rename path(finder) submodule to services
