+++ raiden-ts/src/index.ts
@@ -5,7 +5,7 @@ export { RaidenEvent, RaidenAction } from './actions';
 export { ShutdownReason } from './constants';
 export { RaidenTransfer, RaidenTransferStatus } from './transfers/state';
 export { ChannelState, RaidenChannel, RaidenChannels } from './channels/state';
-export { RaidenPaths, RaidenPFS } from './path/types';
+export { RaidenPaths, RaidenPFS } from './services/types';
 export { RaidenConfig } from './config';
 export * from './types';
 export * from './utils/types';
